[ADD] Basic AI feature (opponent)
[ADD] Player Manager System (with Input Handler)
[ADD] Some tags components to prefabs
[UPDATE] UnitSelected & UnitSelectable to Selected & Selectable component
[FIX] Scale of selection circle for buildings
[REFACTOR] Some files (variables name, serializefield...)

+++ Assets/Scripts/StartMenuPresenter.cs
@@ -31,7 +31,7 @@ public class StartMenuPresenter : MonoBehaviour
 
         var gameManager = GameObject.Find("GameManager").GetComponent<GameManager>();
         gameManager.SelectedRace = selectedRace;
-        gameManager.DifficultyLevel = difficultyLevel;
+        gameManager.DifficultyLevel = (int)difficultyLevel;
         gameManager.PlayerName = playerName; // Stocker le nom du joueur
 
         // Changer de scène
